native: crude hack to build CPM/65 code
If the cpm65 source code is installed in the same directory is this repo,
then the default Makefile will detect it and set CMake variable CPM65_PATH
accordingly. With this variable set, a shell script in src/65c02/cpm will
then assemble all *.asm files there to src/bin/cpm/0/*.com.

More intended to be a temporary solution, let's see how long it stays in
here.

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>