65c02: added function to run external make projects
The function "runmake" adds the option to generate a single binary from a
subdirectory within "src/65c02". This is more like a crutch, because the
"Makefile" needs to be adapted to the build environment. The "Makefile"
however needs to honor the variable "OUTDIR" pointing to the directory
where both the intermediate ".o" files and the final results should be
written to.

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>