addition of instant assembler by Alan Cashin
Instant 6502 Assembler was originally written for the KIM-1 and
reconstructed using the acme assembler. This is a port to the ca65 as
assembler and Sorbus as platform. The port is elegant in a way that the
routines called by instant assembler in the KIM-1 have been reimplemented
for the Sorbus.

So by flipping a switch in the source code, the KIM-1 version can still
be assembled. (Theoretically, since it was not tried.)

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>