fixed cpmfs save bug and preparing next release
cpmfs had a bug where filed bigger 16k were truncated to size of final
chunk.

On Deadline 2025 (https://demoparty.berlin), there will be a demo released
using this platform with FB32X32_LEDs matrix. Following this, there will
be a release including the demo. Kernel Version bumped to 0.6 for this.

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>